requirements.txt
scenedetect[opencv]==0.6.4
opencv-python==4.10.0.84
openai-whisper==20231117
torch==2.2.2
janome==0.5.0
pyyaml==6.0.2
tqdm==4.66.5
rich==13.7.1
numpy==1.26.4
